feat(core): add suport to computed property names for ES5
